+++ src/tqec/_cli/subcommands/dae2circuits.py
@@ -2,13 +2,16 @@ from __future__ import annotations
 
 import argparse
 from pathlib import Path
+
 from typing_extensions import override
 
+from tqec import annotate_detectors_automatically
 from tqec._cli.subcommands.base import TQECSubCommand
 from tqec._cli.subcommands.dae2observables import save_correlation_surfaces_to
-from tqec.computation.block_graph import BlockGraph
 from tqec.compile.compile import compile_block_graph
-from tqec import annotate_detectors_automatically
+from tqec.compile.specs.library.css import CSS_SPEC_RULES
+from tqec.compile.substitute import DEFAULT_SUBSTITUTION_RULES
+from tqec.computation.block_graph import BlockGraph
 
 
 class Dae2CircuitsTQECSubCommand(TQECSubCommand):
@@ -99,6 +102,8 @@ class Dae2CircuitsTQECSubCommand(TQECSubCommand):
         circuits_out_dir.mkdir(exist_ok=True)
         compiled_graph = compile_block_graph(
             block_graph,
+            spec_rules=CSS_SPEC_RULES,
+            substitute_rules=DEFAULT_SUBSTITUTION_RULES,
             observables=[abstract_observables[i] for i in obs_indices],
         )
         ks: list[int] = args.k
