+++ src/tqec/compile/__init__.py
@@ -1,4 +1,4 @@
 from .block import CompiledBlock
 from .compile import compile_block_graph
-from .specs import DEFAULT_SPEC_RULES, CubeSpec, SpecRule
+from .specs import CubeSpec, SpecRule
 from .substitute import DEFAULT_SUBSTITUTION_RULES, SubstitutionKey, SubstitutionRule
